Use synchronous exception handling
This is all nice and well, but the API calls themselves are asynchronous
and therefore may fail after they've been kicked off. Empirical
experience shows that we may get an error back from Pinboard ~5s after
the GET request was made. Some sort of async exception handling needs to
be brought in to deal with this scenario.